ScreenShare/requirements.txt
flask
flask-sock
mss
pillow
pyaudiowpatch
